creating classes for each structure
