+++ yadi/dss/line.py
@@ -1,0 +1,39 @@
+
+import numpy as np
+import pandas as pd
+import yadi.yadi.dss.bus as bus 
+import os
+
+
+class DSS_Line(bus.DSS_Bus):
+
+    def __init__(self, redirects, precompile, verbose=False):
+        """"
+        Class for handling lines in OpenDSS.
+
+        """
+
+        super().__init__(redirects, redirects, precompile)
+
+    def get_lineEAmps(self):
+        "Method to extract line emergency amps"
+        lines = self.dss.Lines.AllNames()
+        thermalLimitDict = dict()
+        for line in lines:
+            self.dss.Lines.Name(line)
+            thermalLimitDict[line] = self.dss.Lines.NormAmps()
+        return pd.Series(thermalLimitDict)
+
+    def get_linePerPhaseEAmps(self):
+        "Method to extract line emergency amps"
+        lines = self.dss.Lines.AllNames()
+        thermalLimitDict = dict()
+        for line in lines:
+            self.dss.Lines.Name(line)
+            phases = self.dss.Lines.Phases()
+            if phases > 1:
+                for ph in range(phases):
+                    thermalLimitDict[line + f".{ph + 1}"] = self.dss.Lines.NormAmps()
+            else:
+                thermalLimitDict[line] = self.dss.Lines.NormAmps()
+        return pd.Series(thermalLimitDict)
